[Bronze V] Title: 이상한 기호, Time: 52 ms, Memory: 5504 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -3,14 +3,14 @@ class Program
     static void Main(string[] args)
     {
         string[] inuputNum = Console.ReadLine().Split();
-        int a = int.Parse(inuputNum[0]);
-        int b = int.Parse(inuputNum[1]);
+        long a = int.Parse(inuputNum[0]);
+        long b = int.Parse(inuputNum[1]);
         
         Console.Write(Calculate(a, b));
         
     }
     
-    static int Calculate(int a, int b)
+    static long Calculate(long a, long b)
     {
         return (a + b) * (a - b);
     }
